Made a script that goes on the Placement Tiles which holds the number that tile is in the List of tiles in the grid. tileSet allows for only changing the tile if it has been left and come back to. Also uses TilePlacer's onTileClicked() so we know what tile to play with.
Modified the UI tiles for this project and added a non-accessible void tile for when you press the Void Fill button.
Renamed FloorUI to Floor and WallUI to Wall so when we save the CSV, it doesn't have the 'UI' part of the name.

For CameraBehaviour.cs:
Added logic so that when you are fully zoomed out, the camera will move to grid centre so it looks pretty once again.
Also added logic to check if the Level Name Input Field isFocused so we know not to move the camera when we press W or A or S or D.

For TilePlacer.cs:
Added logic for placing tiles on the grid, finally!
Added logic so that only one player tile can ever exist on the board.
Added a List which will hold all the spawned-in tiles so that we have reference to them all in a ordered way for later.
Got rid of scaling code, it worked but it was gross and not necessary, it needed to be fixed.
Added logic for saving to CSV using a StringBuilder and doing smart things with that.
Added logic for If you try to save with a name that is already the name of another CSV file, asking if you want to overwrite that CSV file or if it was a mistake you clicking 'Save'.
Added a check to see if every Placement Tile has been replaced with a different tile (also prompting to click the Void Fill button to fill the rest of the grid with Void tiles).
Yes you can press VoidFIll and it will replace the left over Placement Tiles with Void tile.

For UITile.cs:
Changed swapping tiles from having to press LeftCtrl to LeftShift as with the next addition, menus would pop up instead of tiles being selected to be swapped.

For ChooseTile.cs:
Added logic for swapping UI tiles with their keybindings.
Needed to change pressing LeftCtrl to LeftShift as explained above.